# docs/swagger-admin.yaml
openapi: 3.0.3
info:
  title: Sorvetudos API - Administração
  description: API destinada ao gerenciamento administrativo da Sorvetudos
  termsOfService: http://swagger.io/terms/
  version: 1.0.0
...
servers:
  - url: http://localhost:8080
paths:
...
  /v1/sorvetudos/admin/usuarios:
    post:
      tags:
        - usuarios
      summary: Adiciona um novo usuario
      description: Adiciona um novo usuario no banco de dados
      requestBody:
        description: JSON contendo os dados do novo usuario a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
        required: true
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Invalid input
        '415':
          description: Error Content type
        '500':
          description: Error Internal Server (Model or Controller)
    get:
      tags:
        - usuarios
      summary: Lista todos os usuarios
      description: Listagem de todos os usuarios registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (Model or Controller)
  /v1/sorvetudos/admin/usuarios/{id}:
    get:
      tags:
        - usuarios
      summary: Busca um usuario pelo ID
      description: Retorna um usuario pelo ID
      parameters:
        - name: id
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Invalid ID supplied
        '404':
          description: Not Found
    put:
      tags:
        - usuarios
      summary: Update an existing usuario
      description: Update an existing usuario by Id
      parameters:
        - name: id
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update an existent usuario in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Invalid ID supplied
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (Model or Controller)
    delete:
      tags:
        - usuarios
      summary: Delete usuario
      description: Excluir um usuario
      parameters:
        - name: id
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Invalid ID supplied
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (Model or Controller)
  /v1/sorvetudos/admin/produtos:
    post:
      tags:
        - produtos
      summary: Adiciona um novo produto
      description: Adiciona um novo produto no banco de dados
      requestBody:
        description: Form-Data contendo os dados do novo produto a ser inserido no sistema
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Produto'
        required: true
      responses:
        '201':
          description: Successful operation
          content:
            multipart/form-data:
              schema:
                $ref: '#/components/schemas/Produto'
        '400':
          description: Invalid input
        '415':
          description: Error Content type
        '500':
          description: Error Internal Server (Model or Controller)
    get:
      tags:
        - produtos
      summary: Lista todos os produtos
      description: Listagem de todos os produtos registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Produto'
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (Model or Controller)
  /v1/sorvetudos/admin/produtos/{id}:
    get:
      tags:
        - produtos
      summary: Busca um produto pelo ID
      description: Retorna um produto pelo ID
      parameters:
        - name: id
          in: path
          description: ID do produto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '400':
          description: Invalid ID supplied
        '404':
          description: Not Found
    put:
      post:
      tags:
        - produtos
      summary: Atualiza um produto
      description: Adiciona um novo produto no banco de dados
      parameters:
        - name: id
          in: path
          description: ID do produto
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Form-Data contendo os dados do produto a ser atualizado no sistema
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Produto'
        required: true
      responses:
        '201':
          description: Successful operation
          content:
            multipart/form-data:
              schema:
                $ref: '#/components/schemas/Produto'
        '400':
          description: Invalid input
        '415':
          description: Error Content type
        '500':
          description: Error Internal Server (Model or Controller)
    delete:
      tags:
        - produtos
      summary: Delete produto
      description: Excluir um produto
      parameters:
        - name: id
          in: path
          description: ID do produto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '400':
          description: Invalid ID supplied
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (Model or Controller)
...
  /v1/sorvetudos/admin/lotes:
    post:
      tags:
        - lotes
      summary: Adiciona um novo lote
      description: Adiciona um novo lote no banco de dados
      requestBody:
        description: JSON contendo os dados do novo lote a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lote'
        required: true
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lote'
        '400':
          description: Invalid input
        '415':
          description: Error Content type
        '500':
          description: Error Internal Server (Model or Controller)
    get:
      tags:
        - lotes
      summary: Lista todos os lotes
      description: Listagem de todos os lotes registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lote'
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (Model or Controller)
  /v1/sorvetudos/admin/lotes/{id}:
    get:
      tags:
        - lotes
      summary: Busca um lote pelo ID
      description: Retorna um lote pelo ID
      parameters:
        - name: id
          in: path
          description: ID do lote
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lote'
        '400':
          description: Invalid ID supplied
        '404':
          description: Not Found
    put:
      tags:
        - lotes
      summary: Update an existing lote
      description: Update an existing lote by Id
      parameters:
        - name: id
          in: path
          description: ID do lote
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update an existent lote in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lote'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lote'
        '400':
          description: Invalid ID supplied
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (Model or Controller)
    delete:
      tags:
        - lotes
      summary: Delete lote
      description: Excluir um lote
      parameters:
        - name: id
          in: path
          description: ID do lote
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lote'
        '400':
          description: Invalid ID supplied
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (Model or Controller)
  /v1/sorvetudos/admin/tamanhos:
    post:
      tags:
        - tamanhos
      summary: Adiciona um novo tamanho
      description: Adiciona um novo tamanho no banco de dados
      requestBody:
        description: JSON contendo os dados do novo tamanho a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tamanho'
        required: true
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tamanho'
        '400':
          description: Invalid input
        '415':
          description: Error Content type
        '500':
          description: Error Internal Server (Model or Controller)
    get:
      tags:
        - tamanhos
      summary: Lista todos os tamanhos
      description: Listagem de todos os tamanhos registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tamanho'
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (Model or Controller)
  /v1/sorvetudos/admin/tamanhos/{id}:
    get:
      tags:
        - tamanhos
      summary: Busca um tamanho pelo ID
      description: Retorna um tamanho pelo ID
      parameters:
        - name: id
          in: path
          description: ID do tamanho
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tamanho'
        '400':
          description: Invalid ID supplied
        '404':
          description: Not Found
    put:
      tags:
        - tamanhos
      summary: Update an existing tamanho
      description: Update an existing tamanho by Id
      parameters:
        - name: id
          in: path
          description: ID do tamanho
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update an existent tamanho in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tamanho'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tamanho'
        '400':
          description: Invalid ID supplied
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (Model or Controller)
    delete:
      tags:
        - tamanhos
      summary: Delete tamanho
      description: Excluir um tamanho
      parameters:
        - name: id
          in: path
          description: ID do tamanho
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tamanho'
        '400':
          description: Invalid ID supplied
        '404':
          description: Not Found
        '500':
          description: Error Internal Server (Model or Controller)
...
components:
  schemas:
...
    Usuario:
      required:
        - nome
        - email
        - senha
        - nivel_de_acesso
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        senha:
          type: string
        nivel_de_acesso:
          type: boolean
    Produto:
      required:
        - nome
        - descricao
        - preco
        - status
        - tamanho
        - img
        - categorias
        - sabores
        - ingredientes
        - tags
        - promocoes
        - lote
        - tamanhos
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        descricao:
          type: string
        preco:
          type: number
        status:
          type: boolean
        tamanho:
          type: string
        img:
          type: file
        categorias:
          type: array
          items:
            type: integer
        sabores:
          type: array
          items:
            type: integer
        ingredientes:
          type: array
          items:
            type: integer
        tags:
          type: array
          items:
            type: integer
        promocoes:
          type: array
          items:
            type: integer
        lote:
          type: array
          items:
            type: integer
        tamanhos:
          type: array
          items:
            type: integer
...
    Lote:
      required:
        - data_fabricacao
        - data_validade
        - quantidade
      type: object
      properties:
        id:
          type: integer
        data_fabricacao:
          type: string
        data_validade:
          type: string
        quantidade:
          type: integer
    Tamanho:
      required:
        - tamanho
      type: object
      properties:
        id:
          type: integer
        tamanho:
          type: string